Expose ColormapNameComboBox in silx.gui.widgets
